wip - extracted filteredCardsQuery, but need to type return type
